[feat] enhance scene pipeline and events, add scene & prefab editing capabilities (#211)
* enhance scene pipeline and events, add scene & prefab editing capabilities

# Conflicts:
#	src/core/scene/scene-process/service/scene.ts

* refine code

* refine code

* refine code && add prefab proxy

# Conflicts:
#	src/api/scene/node-schema.ts

* refine code

* add scene configs

* refine code

* refine code

* refine code

* refine code

* revert remove cocos.config.json

* refine schema code

* refine code

* refine code

* refine code

* refine code

* refine code

* improve editor reload

* refine code

* refine code

* refine code

* fix node manager event multiple registrations

* add serialize log

* Revert "add serialize log"

This reverts commit 93998434442f64daa5128c9c801b94c319e45e40.

* Reapply "add serialize log"

This reverts commit c0f05a4883a9bfa5185eb2c050e4d866eda7120f.

* fix node event

* refine code

* refine code

* Revert "refine code"

This reverts commit cd9f8db8a8cb31a83cce69d0adb131ae7369ac97.

* Revert "Reapply "add serialize log""

This reverts commit 55b64a021185ed46a9b489a850cd34c75baad59e.

* refine rpc code

* refine code

* refine process-rpc

* refine process-rpc

* simplify process-rpc code

* refine code, add rpc test

* refine code

* refine code

* refine rpc

* fix open scene error

* refine code

* refine code

* refine code

* rm asset events

+++ src/core/scene/test/scene.test.ts
@@ -1,27 +1,31 @@
 import * as fse from 'fs-extra';
 import { SceneTestEnv } from './scene-test-env';
 
+function removeCache() {
+    if (fse.existsSync(SceneTestEnv.cacheDirectory)) {
+        fse.removeSync(SceneTestEnv.cacheDirectory);
+        fse.removeSync(SceneTestEnv.cacheDirectory + '.meta');
+        console.log('删除场景测试目录:', SceneTestEnv.cacheDirectory);
+    }
+}
+
 beforeAll(async () => {
+    removeCache();
     fse.ensureDirSync(SceneTestEnv.cacheDirectory);
     console.log('创建场景测试目录:', SceneTestEnv.cacheDirectory);
     const TestUtils = await import('../../test/global-setup');
     await TestUtils.globalSetup();
 });
 
-afterAll(() => {
-    try {
-        fse.removeSync(SceneTestEnv.cacheDirectory);
-        fse.removeSync(SceneTestEnv.cacheDirectory + '.meta');
-        console.log('删除场景测试目录:', SceneTestEnv.cacheDirectory);
-    } catch (e) { }
+afterAll(async () => {
+    removeCache();
 });
 
-import './scene-proxy.testcase';
+import './editor-proxy-scene.testcase';
+import './editor-proxy-prefab.testcase';
 import './node-proxy.testcase';
 import './component-proxy.testcase';
+import './prefab-proxy.testcase';
 import './script-proxy.testcase';
 import './engine-proxy.testcase';
 import './scene-exit.testcase';
-
-
-
